requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
onnxruntime==1.16.1
pillow==10.1.0
numpy==1.24.3
pydantic==2.4.2
python-dotenv==1.0.0
gunicorn==21.2.0 